backend/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
python-multipart==0.0.9
requests==2.32.3
google-api-python-client==2.141.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
